chore(fibonacci/consolidate): tests and questions
All done.
